pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "livekit-agents[google,silero,turn-detector]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "google-genai==1.64.0",
    "python-dotenv>=1.2.1",
]
